auth_service/requirements.txt
django>=4.0,<5.0
djangorestframework>=3.14.0,<4.0
psycopg2-binary>=2.9.5,<3.0
django-cors-headers>=4.0.0,<5.0
djangorestframework-simplejwt>=5.2.0,<6.0
requests>=2.28.0,<3.0
python-consul>=1.1.0,<2.0
